# Improve grace hash join performance by rerange the right join table by keys