--- a/src/Processors/Transforms/JoiningTransform.cpp
+++ b/src/Processors/Transforms/JoiningTransform.cpp
@@ -351,12 +351,11 @@ void FillingRightJoinSideTransform::work()
         stop_reading = !join->addBlockToJoin(block);
     }
 
-    if (input.isFinished())
+    if (input.isFinished() && !join->supportParallelJoin())
         join->tryRerangeRightTableData();
 
     set_totals = for_totals;
 }
-
 
 DelayedJoinedBlocksWorkerTransform::DelayedJoinedBlocksWorkerTransform(
     Block output_header_,